+++ android/app/src/main/java/edu/illinois/rokwire/maps/MapView.java
@@ -19,6 +19,8 @@ package edu.illinois.rokwire.maps;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
+import android.content.res.Resources;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewParent;
@@ -32,6 +34,7 @@ import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.model.CameraPosition;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
+import com.google.android.gms.maps.model.MapStyleOptions;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.gson.Gson;
@@ -47,6 +50,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import androidx.annotation.NonNull;
 import edu.illinois.rokwire.Constants;
 import edu.illinois.rokwire.MainActivity;
 import edu.illinois.rokwire.R;
@@ -60,7 +64,11 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
     private Activity activity;
     private com.google.android.gms.maps.MapView googleMapView;
     private GoogleMap googleMap;
-    private List<Object> explores;
+    private MapStyleOptions mapStyleOptions;
+
+    private ArrayList<Object> explores;
+    private HashMap exploreOptions;
+    private List<Object> displayExplores;
     private List<Marker> markers;
 
     private IconGenerator iconGenerator;
@@ -106,13 +114,15 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
         googleMapView.layout(0, 0, r, b);
         if (!mapLayoutPassed) {
             mapLayoutPassed = true;
-            showExploresOnMap();
+            acknowledgeExplores();
         }
     }
 
     private void init() {
+        initValuesFromArguments();
         initMarkerView();
         initMapView();
+        initMapStyleOptions();
     }
 
     private void initMapView() {
@@ -124,7 +134,6 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
     }
 
     private void initMarkerView() {
-        acknowledgeValuesFromArgs();
         iconGenerator = new IconGenerator(activity);
         iconGenerator.setBackground(activity.getDrawable(R.color.transparent));
         LayoutInflater inflater = (activity != null) ? (LayoutInflater) activity.getSystemService(Context.LAYOUT_INFLATER_SERVICE) : null;
@@ -132,20 +141,30 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
         markerGroupLayoutView = (inflater != null) ? inflater.inflate(R.layout.marker_group_layout, null) : null;
     }
 
+    private void initMapStyleOptions() {
+        try {
+            mapStyleOptions = MapStyleOptions.loadRawResourceStyle(context, R.raw.mapstyle_nopoi);
+        } catch (Resources.NotFoundException e) {
+            Log.e("MapView", "Failed to load map style options from resources. Stacktrace:");
+            e.printStackTrace();
+        }
+    }
+
     @Override
-    public void onMapReady(GoogleMap map) {
+    public void onMapReady(@NonNull GoogleMap map) {
         onResume();
         googleMap = map;
         enableMyLocation(enableLocationValue);
+        googleMap.setMapStyle(mapStyleOptions);
         googleMap.setOnMarkerClickListener(this::onMarkerClicked);
         googleMap.setOnMapClickListener(this::onMapClick);
-        googleMap.setOnCameraIdleListener(this::updateMarkers);
+        googleMap.setOnCameraIdleListener(this::onCameraIdle);
         googleMap.moveCamera(CameraUpdateFactory.newCameraPosition(CameraPosition.fromLatLngZoom(Constants.DEFAULT_INITIAL_CAMERA_POSITION, Constants.DEFAULT_CAMERA_ZOOM)));
         showExploresOnMap();
         relocateMyLocationButton();
     }
 
-    private void acknowledgeValuesFromArgs() {
+    private void initValuesFromArguments() {
         boolean myLocationEnabled = false;
         if (args instanceof Map) {
             Map<String, Object> jsonArgs = (Map) args;
@@ -159,32 +178,11 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
         this.enableLocationValue = myLocationEnabled;
     }
 
-    private void moveCameraToSpecificPosition() {
-        CameraUpdate cameraUpdate;
-        if ((markers == null) || markers.isEmpty()) {
-            cameraUpdate = CameraUpdateFactory.newLatLngZoom(Constants.DEFAULT_INITIAL_CAMERA_POSITION, Constants.DEFAULT_CAMERA_ZOOM);
-        }
-        else if (markers.size() == 1) {
-            cameraUpdate = CameraUpdateFactory.newLatLngZoom(markers.get(0).getPosition(), Constants.DEFAULT_CAMERA_ZOOM);
-        }
-        else {
-            LatLngBounds.Builder builder = new LatLngBounds.Builder();
-            for (Marker marker : markers) {
-                builder.include(marker.getPosition());
-            }
-            LatLngBounds bounds = builder.build();
-            int width = getResources().getDisplayMetrics().widthPixels;
-            int height = getResources().getDisplayMetrics().heightPixels;
-            int padding = 150; // offset from edges of the map in pixels
-            cameraUpdate = CameraUpdateFactory.newLatLngBounds(bounds, width, height, padding);
-        }
-        googleMap.animateCamera(cameraUpdate);
-    }
-
     public void applyExplores(ArrayList explores, HashMap options) {
-        this.explores = buildExplores(explores, options);
+        this.explores = explores;
+        this.exploreOptions = options;
         if (mapLayoutPassed) {
-            showExploresOnMap();
+            acknowledgeExplores();
         }
     }
 
@@ -206,14 +204,41 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
         }
     }
 
-    private List<Object> buildExplores(ArrayList rawExplores, HashMap options) {
+    private void acknowledgeExplores() {
+        if ((explores != null) && mapLayoutPassed) {
+            LatLngBounds bounds = getBoundsOfExplores(explores);
+            if (bounds != null) {
+                final int cameraPadding = 150;
+                CameraUpdate cameraUpdate = CameraUpdateFactory.newLatLngBounds(bounds, cameraPadding);
+                googleMap.moveCamera(cameraUpdate);
+                double thresholdDistance;
+                Object exploreLocationThresholdParam = (exploreOptions != null) ? exploreOptions.get("LocationThresoldDistance") : null;
+                if (exploreLocationThresholdParam instanceof Double) {
+                    thresholdDistance = (Double) exploreLocationThresholdParam;
+                } else {
+                    thresholdDistance = getThresholdDistance(googleMap.getCameraPosition().zoom);
+                }
+                buildDisplayExploresForThresholdDistance(thresholdDistance);
+            }
+        }
+    }
+
+    private void buildDisplayExplores() {
+        if (mapLayoutPassed) {
+            Object exploreLocationThresholdParam = (exploreOptions != null) ? exploreOptions.get("LocationThresoldDistance") : null;
+            double thresholdDistance = (exploreLocationThresholdParam instanceof Double) ? (Double) exploreLocationThresholdParam : getAutomaticThresholdDistance();
+            buildDisplayExploresForThresholdDistance(thresholdDistance);
+        }
+    }
+
+    private void buildDisplayExploresForThresholdDistance(double thresholdDistance) {
+        displayExplores = buildExplores(explores, thresholdDistance);
+        showExploresOnMap();
+    }
+
+    private List<Object> buildExplores(ArrayList rawExplores, double thresholdDistance) {
         if (rawExplores == null || rawExplores.size() == 0) {
             return null;
-        }
-        Object exploreLocationThresholdParam = (options != null) ? options.get("LocationThresoldDistance") : null;
-        double exploreLocationThresholdDistance = Constants.EXPLORE_LOCATION_THRESHOLD_DISTANCE;
-        if (exploreLocationThresholdParam instanceof Double) {
-            exploreLocationThresholdDistance = (Double) exploreLocationThresholdParam;
         }
         List<ArrayList<HashMap>> mappedExploreGroups = new ArrayList<>();
         int rawExploresCount = rawExplores.size();
@@ -232,7 +257,7 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
                             Integer mappedExploreFloor = Utils.Explore.optLocationFloor(mappedExplore);
                             boolean sameFloor = (exploreFloor == null && mappedExploreFloor == null) ||
                                     ((exploreFloor != null && mappedExploreFloor != null) && exploreFloor.equals(mappedExploreFloor));
-                            if ((distance != null) && (distance < exploreLocationThresholdDistance) && sameFloor) {
+                            if ((distance != null) && (distance <= thresholdDistance) && sameFloor) {
                                 mappedExploreGroup.add(explore);
                                 exploreMapped = true;
                                 break;
@@ -266,9 +291,9 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
             return;
         }
         clearMarkers();
-        if (explores != null && explores.size() > 0) {
+        if (displayExplores != null && displayExplores.size() > 0) {
             markers = new ArrayList<>();
-            for (Object explore : explores) {
+            for (Object explore : displayExplores) {
                 MarkerOptions markerOptions = Utils.Explore.constructMarkerOptions(getContext(), explore, markerLayoutView, markerGroupLayoutView, iconGenerator);
                 if (markerOptions != null) {
                     Marker marker = googleMap.addMarker(markerOptions);
@@ -279,7 +304,6 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
             }
         }
         updateMarkers();
-        moveCameraToSpecificPosition();
     }
 
     private synchronized void clearMarkers() {
@@ -294,16 +318,28 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
 
     private void updateMarkers() {
         float currentCameraZoom = googleMap.getCameraPosition().zoom;
-        boolean updateMarkerInfo = (currentCameraZoom != cameraZoom);
-        if (markers != null && !markers.isEmpty()) {
-            for (Marker marker : markers) {
-                if (updateMarkerInfo) {
+        boolean updateMarkerInfo = Utils.Explore.shouldUpdateMarkerView(currentCameraZoom, cameraZoom);
+        if (updateMarkerInfo) {
+            if (markers != null && !markers.isEmpty()) {
+                for (Marker marker : markers) {
                     boolean singleExploreMarker = Utils.Explore.optSingleExploreMarker(marker);
                     Utils.Explore.updateCustomMarkerAppearance(getContext(), marker, singleExploreMarker, currentCameraZoom, cameraZoom, markerLayoutView, markerGroupLayoutView, iconGenerator);
                 }
             }
         }
         cameraZoom = currentCameraZoom;
+    }
+
+    private void updateMapStyle() {
+        boolean hideBuildingLabelsValue = false;
+        Object hideBuildingsLabelParam = (exploreOptions != null) ? exploreOptions.get("HideBuildingLabels") : null;
+        if (hideBuildingsLabelParam instanceof Boolean) {
+            hideBuildingLabelsValue = (Boolean) hideBuildingsLabelParam;
+        }
+        if (hideBuildingLabelsValue) {
+            MapStyleOptions mapStyle = (cameraZoom >= Constants.MAP_NO_POI_THRESHOLD_ZOOM) ? mapStyleOptions : null;
+            googleMap.setMapStyle(mapStyle);
+        }
     }
 
     private boolean onMarkerClicked(Marker marker) {
@@ -346,6 +382,20 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
         MainActivity.invokeFlutterMethod("map.explore.clear", methodArguments);
     }
 
+    private void onCameraIdle() {
+        CameraPosition cameraPosition = googleMap.getCameraPosition();
+        float zoomDelta = Math.abs(cameraZoom - cameraPosition.zoom);
+        boolean zoomDeltaPassed = (zoomDelta > Constants.MAP_THRESHOLD_ZOOM_UPDATE_STEP);
+        // Prevent building explores (respectively markers) for zoom level that is above our max zoom level
+        boolean maxZoomLevelNotReached = (cameraPosition.zoom < Constants.MAP_MAX_ZOOM_LEVEL_FOR_THRESHOLD);
+        if (zoomDeltaPassed && maxZoomLevelNotReached) {
+            buildDisplayExplores();
+        } else {
+            updateMarkers();
+        }
+        updateMapStyle();
+    }
+
     private void relocateMyLocationButton() {
         if (googleMapView == null) {
             return;
@@ -367,5 +417,44 @@ public class MapView extends FrameLayout implements OnMapReadyCallback {
         rlp.addRule(RelativeLayout.ALIGN_PARENT_TOP, 0);
         rlp.addRule(RelativeLayout.ALIGN_PARENT_BOTTOM, RelativeLayout.TRUE);
         rlp.setMargins(0, 0, 30, 30);
+    }
+
+    private LatLngBounds getBoundsOfExplores(ArrayList<Object> explores) {
+        LatLngBounds.Builder boundsBuilder = null;
+        if ((explores != null) && !explores.isEmpty()) {
+            for (Object explore : explores) {
+                if (explore instanceof HashMap) {
+                    LatLng exploreLatLng = Utils.Explore.optLocationLatLng((HashMap) explore);
+                    if (exploreLatLng != null) {
+                        if (boundsBuilder == null) {
+                            boundsBuilder = new LatLngBounds.Builder();
+                        }
+                        boundsBuilder.include(exploreLatLng);
+                    }
+                }
+            }
+        }
+        return (boundsBuilder != null) ? boundsBuilder.build() : null;
+    }
+
+    private double getAutomaticThresholdDistance() {
+        return getThresholdDistance(googleMap.getCameraPosition().zoom);
+    }
+
+    private float getThresholdDistance(float zoom) {
+        final float[] thresholdDistanceByZoom = {
+            1000000, 800000, 600000, 200000, 100000,    // zoom 0 - 4
+            100000,  80000,  60000,  20000,  10000,     // zoom 5 - 9
+            5000,    1000,   500,    200,    100,       // zoom 10 - 14
+            50,      0                                  // zoom 15 - 16 (max zoom level)
+        };
+
+        int zoomIndex = Math.round(zoom);
+        if ((zoomIndex >= 0) && (zoomIndex < thresholdDistanceByZoom.length)) {
+            float zoomDistance = thresholdDistanceByZoom[zoomIndex];
+            float nextZoomDistance = ((zoomIndex + 1) < thresholdDistanceByZoom.length) ? thresholdDistanceByZoom[zoomIndex + 1] : 0;
+            return zoomDistance - (zoom - (float) zoomIndex) * (zoomDistance - nextZoomDistance);
+        }
+        return 0;
     }
 }