+++ lib/ui/groups/GroupCreatePanel.dart
@@ -90,7 +90,8 @@ class _GroupCreatePanelState extends State<GroupCreatePanel> {
     _researchConsentDetailsController.text = _group?.researchConsentDetails ?? '';
     _authManGroupNameController.text = _group?.authManGroupName ?? '';
 
-    _researchRequiresConsentConfirmation = StringUtils.isNotEmpty(_group?.researchConsentStatement);
+    // #2550: we need consent checkbox selected by default
+    _researchRequiresConsentConfirmation = true; // StringUtils.isNotEmpty(_group?.researchConsentStatement)
 
     if (StringUtils.isNotEmpty(_group?.researchConsentStatement)) {
       _researchConsentStatementController.text = _group!.researchConsentStatement!;
@@ -215,7 +216,7 @@ class _GroupCreatePanelState extends State<GroupCreatePanel> {
                         _buildMembershipLayout(),
                       ],),),),
                       
-                      Visibility(visible: _isManagedGroupAdmin && !_isResearchProject, child:
+                      Visibility(visible: !_isResearchProject, child:
                         Padding(padding: EdgeInsets.only(top: 8), child:
                           _buildCanAutojoinLayout(),
                         )
@@ -402,7 +403,7 @@ class _GroupCreatePanelState extends State<GroupCreatePanel> {
     return Container(padding: EdgeInsets.only(left: 16, right: 16, top: 8), child:
       Column(children: [
         _buildSwitch(
-          title: "Requires confirmation",
+          title: "Require participant consent",
           value: _researchRequiresConsentConfirmation,
           onTap: _onTapResearchConfirmation
         ),
@@ -746,7 +747,7 @@ class _GroupCreatePanelState extends State<GroupCreatePanel> {
         padding: EdgeInsets.only(left: 16, right: 16, top: 8),
         child: _buildSwitch(
             title: "Is the research project open?",
-            value: _group?.researchOpen,
+            value: _group?.researchOpen == true,
             onTap: _onTapResearchOpen));
   }